solution_HIBERNATE: Вынес в AbstractUlil общую логику, поднастроил логирование
