+++ src/main/java/jm/task/core/jdbc/util/UtilHibernate.java
@@ -1,31 +1,38 @@
 package jm.task.core.jdbc.util;
 
 import jm.task.core.jdbc.model.User;
-import org.hibernate.Session;
 import org.hibernate.SessionFactory;
-import org.hibernate.boot.model.naming.CamelCaseToUnderscoresNamingStrategy;
 import org.hibernate.cfg.Configuration;
 
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.Properties;
-import java.util.logging.Level;
+import java.util.logging.Logger;
 
 
-public class UtilHibernate {
+public final class UtilHibernate extends AbstractUtil {
 
     private static SessionFactory sessionFactory; // for Hibernate
+    private static final Logger LOGGER = getLogger(UtilHibernate.class);
 
 
     public static SessionFactory getSessionFactory() {
         final String hibernatePropertiesFileName = "hibernateConfig.properties";
-        if (sessionFactory == null) {
+        if (sessionFactory == null || sessionFactory.isClosed()) {
             try (InputStream input = UtilHibernate.class.getClassLoader().getResourceAsStream(hibernatePropertiesFileName)) {
                 if (input == null) {
                     throw new RuntimeException("Не удалось найти файл '" + hibernatePropertiesFileName + "'");
                 }
                 Properties properties = new Properties();
                 properties.load(input);
+
+                properties.forEach((key, value) -> LOGGER.fine(key + " = " + value));
+
+                LOGGER.fine("getConnection(): Properties loaded successfully;"
+                        + " driver = '" + properties.getProperty("hibernate.connection.driver_class") + "';"
+                        + " url = '" + properties.getProperty("hibernate.connection.url") + "';"
+                        + " username = '" + properties.getProperty("hibernate.connection.username") + "';"
+                        + " password = '" + "*".repeat(properties.getProperty("hibernate.connection.password").length()) + "'");
 
                 Configuration configuration = new Configuration();
 
@@ -34,15 +41,15 @@ public class UtilHibernate {
                 // Подключаем 'hibernate.cfg.xml' / так давно не делают, а делают через аннотации, см. ниже
                 configuration.configure();
 
-
                 // Добавляем аннотированный класс до создания SessionFactory
                 configuration.addAnnotatedClass(User.class); // Добавление аннотированных классов должно происходить до создания SessionFactory. // т.е. до buildSessionFactory()
-
 
                 configuration.setProperties(properties);
 
                 // Строим фабрику сессий
                 sessionFactory = configuration.buildSessionFactory();
+
+                LOGGER.fine("getConnection(): getConnection is successfully; ");
             } catch (IOException e) {
                 System.err.println("\n\n\nException:   SessionFactory getSessionFactory\n\n");
                 e.printStackTrace();
