+++ src/main/resources/db_Hibernate_config.properties
@@ -1,4 +1,4 @@
-## PostgreSQL properties
+## PostgreSQL properties // db_Hibernate_config
 db.driver=org.postgresql.Driver
 db.url=jdbc:postgresql://localhost:5432/ITM_task013_TaskHibernate_Task_2_db
 db.username=user_hibernate
